refactor: standardization of layout component files
