--- frontend/src/components/MainLayout/NavButtons/types.ts
@@ -1,11 +1,0 @@
-import type { ReactNode } from "react"
-
-export type NavButtonsProps = {
-  navItems: NavItem[]
-}
-
-export type NavItem = {
-  icon: ReactNode
-  label: string
-  path: string
-}
