--- frontend/src/components/MainLayout/SideBar/types.ts
@@ -1,8 +1,0 @@
-import type { NavButtonsProps } from "../NavButtons/types"
-import type { UserProfileProps } from "../UserProfile/types"
-
-export type SideBarProps = UserProfileProps &
-  NavButtonsProps & {
-    buttonText: string
-    onButtonClick: () => void
-  }
