--- frontend/src/components/MainLayout/UserProfile/types.ts
@@ -1,3 +1,0 @@
-export type UserProfileProps = {
-  userName: string
-}
